Allow :Rg to take ! (bang flag) like :grep
:Rg command internally called :grep! and there was no way to not pass
this ! flag.
This change added -bang option to :Rg, passing through it to internal
:grep invocation. Because it breaks the original behavior, a new option,
"g:rg_jump_to_first" was added to control the behavior.

When g:rg_jump_to_first is false (by default),

 * :Rg executes :grep!
 * :Rg! executes :grep

It is the same as the original behavior.

When g:rg_jump_to_first is true,

 * :Rg executes :grep
 * :Rg! executes :grep!

The ! flag of :Rg is passed to :grep as it is.